requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
groq==0.9.0
requests==2.32.3
gunicorn==22.0.0
python-dotenv==1.0.1
